+++ docs/changes.rst
@@ -4,8 +4,9 @@ Changelog
 1.0 - Unreleased
 ----------------
 
-* ...
-
+* Refactor storage adapter for image scales to be less dependent on the
+  underlying content type.
+  [witsch]
 
 
 1.0a1 - November 10th, 2009
@@ -13,4 +14,3 @@ Changelog
 
 * Initial release
   [wichert]
-